requirements.txt
fastapi
fastapi-mail
pydantic[email]
uvicorn
requests
python-multipart
psycopg
python-jose
passlib[bcrypt]
azure-storage-blob
azure-identity